requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
torchaudio
opencv-python
numpy
pandas
scikit-learn
matplotlib
diffusers
transformers
accelerate
timm
Pillow
requests
google-generativeai